requirements.txt
flask>=3.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
flask-caching>=2.1.0
gunicorn>=21.2.0
python-json-logger>=2.0.7
prometheus-flask-exporter>=0.22.4
sentry-sdk[flask]>=1.39.1